[R] test: test downPins at frame level
