+++ src/main/java/bowling/domain/frame/Frame.java
@@ -1,5 +1,6 @@
 package bowling.domain.frame;
 
+import bowling.domain.pin.DownedPins;
 import bowling.domain.state.State;
 
 public abstract class Frame {
@@ -8,4 +9,9 @@ public abstract class Frame {
     protected Frame(State state) {
         this.state = state;
     }
+
+    public void downPins(DownedPins downedPins) {
+        this.state = state.downPins(downedPins);
+    }
+
 }
